Accommodate recent change to JSC's print utility
On 2021-09-29, the `print` function in JavaScriptCore was modified to
display string values within quotation mark characters [1]. This
interfered with esvu's ability to install the engine, failing with
an error, e.g.

    JavaScriptCore ❯ Testing...
    esvu ✖ Fatal error installing JavaScriptCore AssertionError [ERR_ASSERTION]: Expected values to be strictly equal:

    '"42"' !== '42'

Update the installation script to accept both this behavior and the
prior behavior in order to preserve functionality when installing old
versions of the engine.

[1] https://bugs.webkit.org/show_bug.cgi?id=230931