Support Boa nightly version
You can get nightly releases from https://github.com/boa-dev/boa/releases/tag/nightly
This adds support for installing them when you don't specify a version
for Boa. The filenames are somewhat different and we tag the version
internally with the updated_at string so that different nightlies aren't
considered to be the same version.